shifted database stuff from RiderMap to DBManager
